Info Menu function added to index
There is sample images in images/menu/
The function runs of the class attached to the button and classes on
the <a href and img tags on the menu.

If changing the info section, it animates up based on the menu above
shrinking. So just replace the holding image with a div and anything in
there will shuffle up.